refactor: applied clippy suggestions

+++ testangel/src/action_loader.rs
@@ -69,33 +69,30 @@ pub fn get_action_directory() -> PathBuf {
 pub fn get_actions(engine_list: &Arc<EngineList>) -> ActionMap {
     let mut actions = HashMap::new();
     let action_dir = get_action_directory();
-    if let Ok(exists) = fs::exists(&action_dir) {
-        if !exists {
+    if let Ok(exists) = fs::exists(&action_dir)
+        && !exists {
             fs::create_dir_all(action_dir.clone()).unwrap();
             let mut path = PathBuf::from(&action_dir);
             path.push("example.taaction");
             let _ = fs::write(path, include_str!("demo_action.taaction"));
         }
-    }
     'action_loop: for path in fs::read_dir(action_dir).unwrap() {
         let path = path.unwrap();
         let filename = path.file_name();
-        if let Ok(meta) = path.metadata() {
-            if meta.is_dir() {
+        if let Ok(meta) = path.metadata()
+            && meta.is_dir() {
                 continue;
             }
-        }
 
-        if let Ok(str) = filename.into_string() {
-            if str.ends_with(".taaction") {
+        if let Ok(str) = filename.into_string()
+            && str.ends_with(".taaction") {
                 tracing::debug!("Detected possible action {str}");
                 if let Ok(res) = fs::read_to_string(path.path()) {
-                    if let Ok(versioned_file) = ron::from_str::<VersionedFile>(&res) {
-                        if versioned_file.version() != 3 {
+                    if let Ok(versioned_file) = ron::from_str::<VersionedFile>(&res)
+                        && versioned_file.version() != 3 {
                             tracing::warn!("Action {str} uses an incompatible file version.");
                             continue 'action_loop;
                         }
-                    }
 
                     if let Ok(action) = ron::from_str::<Action>(&res) {
                         // Validate that all instructions are available for this action before loading
@@ -123,7 +120,6 @@ pub fn get_actions(engine_list: &Arc<EngineList>) -> ActionMap {
                     tracing::warn!("Couldn't read action");
                 }
             }
-        }
     }
     ActionMap(actions)
 }
